+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning][].
 
 ## [0.2.3] - 2024-09-25
 
+### Changed
+
+-   Validation of element and column names allowing `_-.` and alphanumeric characters #624
+
 ### Minor
 
 -   Added `clip: bool = False` parameter to `polygon_query()` #670